Change return type of manhatten distance in Vector3D to double

+++ mcapi/src/internal/java/net/flintmc/mcapi/internal/world/math/DefaultVector3D.java
@@ -85,11 +85,11 @@ public class DefaultVector3D implements Vector3D {
    * {@inheritDoc}
    */
   @Override
-  public int manhattanDistance(Vector3D vector) {
+  public double manhattanDistance(Vector3D vector) {
     double absX = Math.abs(vector.getX() - this.getX());
     double absY = Math.abs(vector.getY() - this.getY());
     double absZ = Math.abs(vector.getZ() - this.getZ());
-    return (int) (absX + absY + absZ);
+    return absX + absY + absZ;
   }
 
   /**
